Move the component into shared namespce.
