+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Components/Lyrics/Components/DrawableTimeTag.cs
@@ -14,7 +14,7 @@ using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Utils;
 using osuTK;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricList.Rows.Edit.Components
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Components
 {
     public class DrawableTimeTag : CompositeDrawable, IHasCustomTooltip<TimeTag>
     {
