+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Components/Lyrics/RubyBlueprintContainer.cs
@@ -9,12 +9,12 @@ using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Game.Rulesets.Edit;
 using osu.Game.Rulesets.Karaoke.Edit.ChangeHandlers.Lyrics;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricList.Rows.Edit.Blueprints;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics.Blueprints;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States.Modes;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Screens.Edit.Compose.Components;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricList.Rows.Edit
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Components.Lyrics
 {
     public class RubyBlueprintContainer : TextTagBlueprintContainer<RubyTag>
     {
